+++ metadata.rb
@@ -5,4 +5,3 @@ license          'GNU Public License 3.0'
 description      'Installs/Configures duplicity_ng'
 long_description 'Installs/Configures duplicity_ng'
 version          '0.1.0'
-