+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 0.1.0
 -----
 
-- [Chris Aumann] - Initial release of duplicity\_ng
+- [Chris Aumann] - Initial release of duplicity\_ng (internal)